refactor: local storage keys
