+++ src/components/Modal/NewsletterSubscriptionModal/NewsletterSubscriptionModal.tsx
@@ -10,10 +10,10 @@ import { Modal, ModalProps } from 'decentraland-ui/dist/components/Modal/Modal'
 import isEmail from 'validator/lib/isEmail'
 
 import { Governance } from '../../../clients/Governance'
+import { ANONYMOUS_USR_SUBSCRIPTION, NEWSLETTER_SUBSCRIPTION_KEY } from '../../../front/localStorageKeys'
 import useAsyncTask from '../../../hooks/useAsyncTask'
 import useFormatMessage from '../../../hooks/useFormatMessage'
 import { NewsletterSubscriptionResult } from '../../../shared/types/newsletter'
-import { ANONYMOUS_USR_SUBSCRIPTION, NEWSLETTER_SUBSCRIPTION_KEY } from '../../Banner/Subscription/SubscriptionBanner'
 import Label from '../../Common/Typography/Label'
 import Text from '../../Common/Typography/Text'
 import CheckCloud from '../../Icon/CheckCloud'
